+++ android/ijkplayer/kantv/src/main/java/tv/danmaku/ijk/media/example/content/RecentMediaStorage.java
@@ -22,13 +22,17 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.media.MediaPlayer;
 import android.os.AsyncTask;
 import android.support.v4.content.AsyncTaskLoader;
 import android.text.TextUtils;
 
+import tv.danmaku.ijk.media.player.IjkLog;
+
 public class RecentMediaStorage {
     private Context mAppContext;
 
+    private final static String TAG = RecentMediaStorage.class.getName();
     public RecentMediaStorage(Context context) {
         mAppContext = context.getApplicationContext();
     }
@@ -37,19 +41,38 @@ public class RecentMediaStorage {
         new AsyncTask<String, Void, Void>() {
             @Override
             protected Void doInBackground(String... params) {
+                IjkLog.d(TAG, "params[0] : " + params[0]);
                 saveUrl(params[0]);
                 return null;
             }
         }.execute(url);
     }
 
-    public void saveUrl(String url) {
+    public void saveUrl(String syncUrl) {
         ContentValues cv = new ContentValues();
-        cv.putNull(Entry.COLUMN_NAME_ID);
-        cv.put(Entry.COLUMN_NAME_URL, url);
-        cv.put(Entry.COLUMN_NAME_LAST_ACCESS, System.currentTimeMillis());
-        cv.put(Entry.COLUMN_NAME_NAME, getNameOfUrl(url));
-        save(cv);
+        String url = "unknown";
+        String mediaType = "unknown";
+        try {
+            String[] strArr = syncUrl.split("_KANTV_");
+            if (2 != strArr.length) {
+                IjkLog.w(TAG, "it shouldn't happen here, pls check");
+                return;
+            }
+            url = strArr[0];
+            mediaType = strArr[1];
+            cv.putNull(Entry.COLUMN_NAME_ID);
+            IjkLog.d(TAG,"url: " + url + " mediatype:" + mediaType);
+            cv.put(Entry.COLUMN_NAME_URL, url);
+            cv.put(Entry.COLUMN_NAME_TYPE, mediaType);
+            cv.put(Entry.COLUMN_NAME_LAST_ACCESS, System.currentTimeMillis());
+            cv.put(Entry.COLUMN_NAME_NAME, getNameOfUrl(url));
+            save(cv);
+        } catch (IllegalArgumentException ex) {
+            IjkLog.w(TAG, "can't get url and mediatype" + ex.getMessage());
+            return;
+        } finally {
+           return;
+        }
     }
 
     public void save(ContentValues contentValue) {
@@ -78,6 +101,7 @@ public class RecentMediaStorage {
         public static final String TABLE_NAME = "RecentMedia";
         public static final String COLUMN_NAME_ID = "id";
         public static final String COLUMN_NAME_URL = "url";
+        public static final String COLUMN_NAME_TYPE = "type";
         public static final String COLUMN_NAME_NAME = "name";
         public static final String COLUMN_NAME_LAST_ACCESS = "last_access";
     }
@@ -86,16 +110,18 @@ public class RecentMediaStorage {
             Entry.COLUMN_NAME_ID + " as _id",
             Entry.COLUMN_NAME_ID,
             Entry.COLUMN_NAME_URL,
+            Entry.COLUMN_NAME_TYPE,
             Entry.COLUMN_NAME_NAME,
             Entry.COLUMN_NAME_LAST_ACCESS};
 
     public static class OpenHelper extends SQLiteOpenHelper {
         private static final int DATABASE_VERSION = 1;
-        private static final String DATABASE_NAME = "RecentMedia.db";
+        private static final String DATABASE_NAME = "RecentMedia_NEW.db";
         private static final String SQL_CREATE_ENTRIES =
                 " CREATE TABLE IF NOT EXISTS " + Entry.TABLE_NAME + " (" +
                         Entry.COLUMN_NAME_ID + " INTEGER PRIMARY KEY AUTOINCREMENT, " +
                         Entry.COLUMN_NAME_URL + " VARCHAR UNIQUE, " +
+                        Entry.COLUMN_NAME_TYPE + " VARCHAR UNIQUE, " +
                         Entry.COLUMN_NAME_NAME + " VARCHAR, " +
                         Entry.COLUMN_NAME_LAST_ACCESS + " INTEGER) ";